Extracted acquire, won't work, currently
